1. upgrade apscheduler3.10.4 -> apscheduler3.10.3 (problem - ERROR:  duplicate key value violates unique constraint "apscheduler_job_pkey" DETAIL:  Key (id)=(bbphj8frttd98mehcoko7eer3r-events-23660109-mH0pXjJIyandex.ru(next_conf)) already exists.) 2. rewrite models.py 3. fix small mistakes in README.md

+++ src/wsgi/database/models.py
@@ -14,7 +14,7 @@ class User:
         metadata_obj,
         Column("id", Integer, primary_key=True),
         Column("mm_user_id", String, nullable=False, unique=True),
-        Column("login", String, nullable=False),
+        Column("login", String, nullable=False, unique=True),
         Column("token", String, nullable=False, unique=True),
         Column("timezone", String, nullable=False),
         Column("e_c", Boolean, nullable=False),
@@ -22,70 +22,68 @@ class User:
     )
 
     @classmethod
-    def add_user(cls, **kwargs):
+    def add_user(cls, mm_user_id: str, login: str, token: str, timezone: str, e_c: bool, ch_stat: bool):
         with engine.begin() as connection:
             c = cls.user_account_table.columns
 
             result = connection.execute(
                 cls.user_account_table
                 .insert()
-                .values(kwargs)
+                .values(mm_user_id=mm_user_id, login=login, token=token, timezone=timezone, e_c=e_c, ch_stat=ch_stat)
                 .returning(c)
             )
 
         return result.first()
 
     @classmethod
-    def get_user(cls, **kwargs):
-        col, val = kwargs.popitem()
-
+    def get_user(cls, mm_user_id: str):
         with engine.begin() as connection:
             c = cls.user_account_table.columns
 
             result = connection.execute(
                 cls.user_account_table
                 .select()
-                .where(getattr(c, col) == val)
+                .where(c.mm_user_id == mm_user_id)
             )
 
         return result.first()
 
     @classmethod
-    def remove_user(cls, **kwargs):
-        col, val = kwargs.popitem()
-
+    def remove_user(cls, mm_user_id: str):
         with engine.begin() as connection:
             c = cls.user_account_table.columns
 
             result = connection.execute(
                 cls.user_account_table
                 .delete()
-                .where(getattr(c, col) == val)
+                .where(c.mm_user_id == mm_user_id)
                 .returning(c)
             )
 
         return result.first()
 
     @classmethod
-    def update_user(cls, **kwargs):
+    def update_user(cls, mm_user_id: str, **kwargs):
         with engine.begin() as connection:
             c = cls.user_account_table.columns
 
             result = connection.execute(
                 cls.user_account_table
                 .update()
-                .where(c.mm_user_id == kwargs['mm_user_id'])
+                .where(c.mm_user_id == mm_user_id)
                 .values(kwargs)
                 .returning(c)
             )
 
         return result.first()
 
-    def __init__(self, **kwargs):
-        [setattr(self, col, val) for col, val in kwargs.items()]
-
-    def save(self):
-        return User.add_user(self.mm_user_id, self.login, self.token, self.timezone, self.e_c, self.ch_stat)
+    def __init__(self, mm_user_id, login, token, timezone, e_c, ch_stat):
+        self.mm_user_id = mm_user_id
+        self.login = login
+        self.token = token
+        self.timezone = timezone
+        self.e_c = e_c
+        self.ch_stat = ch_stat
 
 
 class YandexCalendar:
@@ -99,18 +97,18 @@ class YandexCalendar:
             nullable=False
         ),
         Column("cal_id", String, nullable=False, unique=True),
-        Column("sync_token", String, ),
+        Column("sync_token", String, unique=True),
     )
 
     @classmethod
-    def add_one_cal(cls, **kwargs):
+    def add_one_cal(cls, user_id: int, cal_id: str, sync_token: str):
         with engine.begin() as connection:
             c = cls.yandex_calendar_table.columns
 
             result = connection.execute(
                 cls.yandex_calendar_table
                 .insert()
-                .values(kwargs)
+                .values(user_id=user_id, cal_id=cal_id, sync_token=sync_token)
                 .returning(c)
             )
 
@@ -131,38 +129,33 @@ class YandexCalendar:
         return result.all()
 
     @classmethod
-    def get_cal(cls, **kwargs):
-        col, val = kwargs.popitem()
-
+    def get_cal(cls, cal_id: str):
         with engine.begin() as connection:
             c = cls.yandex_calendar_table.columns
 
             result = connection.execute(
                 cls.yandex_calendar_table
                 .select()
-                .where(getattr(c, col) == val)
+                .where(c.cal_id == cal_id)
             )
 
         return result.first()
 
     @classmethod
-    def get_cals(cls, **kwargs):
-        col, val = kwargs.popitem()
-
+    def get_cals(cls, user_id: int):
         with engine.begin() as connection:
             c = cls.yandex_calendar_table.columns
 
             result = connection.execute(
                 cls.yandex_calendar_table
                 .select()
-                .where(getattr(c, col) == val)
+                .where(c.user_id == user_id)
             )
 
         return result.all()
 
     @classmethod
-    def remove_cal(cls, **kwargs):
-        col, val = kwargs.popitem()
+    def remove_cal(cls, cal_id: str):
 
         with engine.begin() as connection:
             c = cls.yandex_calendar_table.columns
@@ -170,15 +163,14 @@ class YandexCalendar:
             result = connection.execute(
                 cls.yandex_calendar_table
                 .delete()
-                .where(getattr(c, col) == val)
+                .where(c.cal_id == cal_id)
                 .returning(c)
             )
 
         return result.first()
 
     @classmethod
-    def remove_cals(cls, **kwargs):
-        col, val = kwargs.popitem()
+    def remove_cals(cls, user_id: int):
 
         with engine.begin() as connection:
             c = cls.yandex_calendar_table.columns
@@ -186,21 +178,21 @@ class YandexCalendar:
             result = connection.execute(
                 cls.yandex_calendar_table
                 .delete()
-                .where(getattr(c, col) == val)
+                .where(c.user_id == user_id)
                 .returning(c)
             )
 
         return result.all()
 
     @classmethod
-    def update_cal(cls, **kwargs):
+    def update_cal(cls, cal_id: str, **kwargs):
         with engine.begin() as connection:
             c = cls.yandex_calendar_table.columns
 
             result = connection.execute(
                 cls.yandex_calendar_table
                 .update()
-                .where(c.cal_id == kwargs.get('cal_id'))
+                .where(c.cal_id == cal_id)
                 .values(kwargs)
                 .returning(c)
             )
@@ -235,59 +227,86 @@ class YandexConference:
     )
 
     @classmethod
-    def get_conference(cls, **kwargs):
-        col, val = kwargs.popitem()
-
+    def get_conference(cls, uid: str):
         with engine.begin() as connection:
             c = cls.yandex_conference_table.columns
 
             result = connection.execute(
                 cls.yandex_conference_table
                 .select()
-                .where(getattr(c, col) == val)
+                .where(c.uid == uid)
             )
 
         return result.first()
 
     @classmethod
-    def add_conference(cls, **kwargs):
+    def add_conference(cls,
+                       cal_id: str,
+                       uid: str,
+                       timezone: str,
+                       dtstart: str,
+                       dtend: str,
+                       summary: str = None,
+                       created: str = None,
+                       last_modified: str = None,
+                       description: str = None,
+                       url_event: str = None,
+                       categories: str = None,
+                       x_telemost_conference: str = None,
+                       organizer: str = None,
+                       attendee: str = None,
+                       location: str = None,
+                       ):
         with engine.begin() as connection:
             c = cls.yandex_conference_table.columns
 
             result = connection.execute(
                 cls.yandex_conference_table
                 .insert()
-                .values(kwargs)
+                .values(cal_id=cal_id,
+                        uid=uid,
+                        timezone=timezone,
+                        dtstart=dtstart,
+                        dtend=dtend,
+                        summary=summary,
+                        created=created,
+                        last_modified=last_modified,
+                        description=description,
+                        url_event=url_event,
+                        categories=categories,
+                        x_telemost_conference=x_telemost_conference,
+                        organizer=organizer,
+                        attendee=attendee,
+                        location=location,
+                        )
                 .returning(c)
             )
 
         return result.first()
 
     @classmethod
-    def remove_conference(cls, **kwargs):
-        col, val = kwargs.popitem()
-
+    def remove_conference(cls, uid: str):
         with engine.begin() as connection:
             c = cls.yandex_conference_table.columns
 
             result = connection.execute(
                 cls.yandex_conference_table
                 .delete()
-                .where(getattr(c, col) == val)
+                .where(c.uid == uid)
                 .returning(c)
             )
 
         return result.first()
 
     @classmethod
-    def update_conference(cls, **kwargs):
+    def update_conference(cls, uid: str, **kwargs):
         with engine.begin() as connection:
             c = cls.yandex_conference_table.columns
 
             result = connection.execute(
                 cls.yandex_conference_table
                 .update()
-                .where(c.uid == kwargs.get('uid'))
+                .where(c.uid == uid)
                 .values(kwargs)
                 .returning(c)
             )
